Player Movement Implemented
Player scale may need to be adjusted.  Most speed variables are tied to Base Swim Speed so changing that should retain other movement properties (drag on the rigidbody might have to be changed as well)

I'm not sure if I like the ascend and descend but we can see how other people think it feels.  Otherwise, I made the camera a bit floaty on purpose, I think it'll really help with the whole underwater feel.